requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.164
botocore==1.12.164
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
en-core-web-sm==2.1.0
entrypoints==0.3
gensim==3.7.3
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.4.3
notebook==5.7.8
numpy==1.16.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
plac==0.9.6
pprint==0.1
preshed==2.0.1
prometheus-client==0.7.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.5.1
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
spacy==2.1.4
srsly==0.0.6
tabulate==0.8.3
terminado==0.8.2
testpath==0.4.2
thinc==7.0.4
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
